Split settings managers and models to regular and CSS oriented
