+++ core/settings/editor-preferences/model.php
@@ -5,7 +5,7 @@ namespace Elementor\Core\Settings\EditorPreferences;
 use Elementor\Controls_Manager;
 use Elementor\Core\Settings\Base\Model as BaseModel;
 
-if( ! defined( 'ABSPATH' ) ) {
+if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly
 }
 
@@ -22,7 +22,7 @@ class Model extends BaseModel {
 	 *
 	 */
 	public function get_name() {
-	 	return 'editor-preferences';
+		return 'editor-preferences';
 	}
 
 	/**
