+++ core/settings/general/manager.php
@@ -4,7 +4,7 @@ namespace Elementor\Core\Settings\General;
 use Elementor\Controls_Manager;
 use Elementor\Core\Files\CSS\Base;
 use Elementor\Core\Files\CSS\Global_CSS;
-use Elementor\Core\Settings\Base\Manager as BaseManager;
+use Elementor\Core\Settings\Base\CSS_Manager;
 use Elementor\Core\Settings\Base\Model as BaseModel;
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -19,7 +19,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @since 1.6.0
  */
-class Manager extends BaseManager {
+class Manager extends CSS_Manager {
 
 	/**
 	 * Lightbox panel tab.
