+++ core/settings/general/model.php
@@ -2,7 +2,7 @@
 namespace Elementor\Core\Settings\General;
 
 use Elementor\Controls_Manager;
-use Elementor\Core\Settings\Base\Model as BaseModel;
+use Elementor\Core\Settings\Base\CSS_Model;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
@@ -16,7 +16,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @since 1.6.0
  */
-class Model extends BaseModel {
+class Model extends CSS_Model {
 
 	/**
 	 * Get model name.
